Background for Popups, fonts scaling improvement,  Settings Popup class
Added a class for solid backgrounds
   > TO DO: add border around it
Added a font and improved font scaling between resolutions by creating .spritefonts with different sizes.
Created the Tickbox.cs class although it is empty
   > TO DO: implement the feature
Added feature list for the settings popup in /Decomp/
Added Settings.cs which is a State for the settings popup.
Created a