Font fix, res scaling fix, OnApply event added to MCB
JX-720 font was set to arial, and not joystix monospaced. Also was too small. JX-1440 font was too large

Fixed position of arrows on MCB so that they are consistent across reses

Fixed resolution scaling

Known Issue: Memory builds up when changing resolution. Implementing IDisposable and disposing textures fixes this but you are unable to get the textures back. Maybe indicate to the GC to remove the persistent references? However, problem is not critical.

Cleaned up using statements.