Changed Base Resolution to 640x360, Fixed Issue with Saveing
Changed base resolution to 640x360 so that there is integer scaling although it is not as effective since I still scale by decimals before upscaling. However, this will be useful when I start adding sprites which will be the correct size so I will not have to scale by floats.

Fixed majors bugs with saving both settings (resolution would not save, may have been due to lines added for debugging) and also saves. Movement Speed was not added to the attributes after exiting the character creation screen and also the item names when converting from names to codes had spaces where the others didn't also fixed logical error where it would read the first item of the inventory and then not stop.