Data Saving Rework (FINALLY)
So, the previous encryption method was, frankly, dog shit. So, I decided to completely redesign the encryption.

The previous method would generate a random string of numbers and then place the actual values in pre-determined and constant indexes. The main issue with this was not only scalability but also my crap implementation. To make it worse, I stored the data in a partially human readable state making it even more annoying to code/debug.

This time, I took more simplistic approach to saving the data. I simply have a map to convert ASCII characters into a string with 64 characters in it. To separate keys and their values their is a special SEPERATOR variable (yes, the typo is intentional). This makes the code so much nicer and the encryption is , essentially, as strong as the previous.  Furthermore, it is much more scalable due to the easy-to-understand system.

I am very pleased I did this.